# Conflicts:
#	Modules/Appraisal/Appraisal/Infrastructure/Repositories/MarketComparableTemplateRepository.cs
#	Modules/Integration/Integration/Application/Features/AppraisalResults/GetAppraisalResult/GetAppraisalResultByCaseKeyEndpoint.cs
#	Modules/Integration/Integration/Application/Features/AppraisalResults/GetAppraisalResult/GetAppraisalResultQuery.cs
#	Modules/Integration/Integration/Application/Features/AppraisalResults/GetAppraisalResult/GetAppraisalResultQueryHandler.cs

+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/CreateMachineryProperty/CreateMachineryPropertyRequest.cs
@@ -40,7 +40,7 @@ public record CreateMachineryPropertyRequest(
     string? Location = null,
     string? ConditionUse = null,
     string? MachineCondition = null,
-    int? MachineAge = null,
+    decimal? MachineAge = null,
     string? MachineEfficiency = null,
     string? MachineTechnology = null,
     string? UsagePurpose = null,
